refactor(icon.service): convert IconService to a pipe for better reusability
The IconService has been refactored to implement the PipeTransform interface, allowing it to be used as a standalone pipe. This change enhances reusability and simplifies the process of transforming icon keys into their corresponding class names. The getIcon method now returns an array of strings to support the pipe transformation.